Add update for comments when commenting
