+++ src/modules/getdata.js
@@ -1,5 +1,5 @@
 const artistName = 'nothing but thieves';
-const token = 'BQBPw54lSFWDUeuPWssiypBXUXo-X9x11OvMuTCpEazvncaV5IhXXBDtlxK5fataXEb5aqprr9nCBK6LvODNdH8vvtbeOF7qZDRz4wGOAZYP5m53cq2E3xjNTi-AhtQFAcnMM-j7jviobJ85nWgZWzw9CeMShUM902QgS56q8ujAohBftPTBPw0hlh4QTsumZuYzxwm-v2OBfDon7A0Q4QSBoclG49sEI023M7Md72UhuWBRqjjGAKeoWK4QL3sCx3KKPHqmihE1dnJ1HueC';
+const token = 'BQBtiy_YL6Wd7qmz4hCTHhqBkcp1iixQEYoc60VuQYJcINzHyiyj0FgvOL75baS7htTbrr_6h5OClop17pYzDwuhzpKIar2I4qKwhbl1yaDZLYE1TYJ80NK1B4EIUfBiTSdyq0yTwii7u9bJjUoZACYltLVOI9zpFHTMdpY9vpBfZg_9S9hZYhGwqP-q6VgMUkcyQzTOBgbcoSbnNPLgCGgA4gOlS2huz2Dt-0tMn8Xxm-YyzMKdpp87WV-7_GbEukFN6ZvbH_-WcDw__T-P';
 
 const getdata = async () => {
   const response = await fetch(`https://api.spotify.com/v1/search?q=${artistName}&type=track`, {
