+++ src/modules/postcomment.js
@@ -1,5 +1,4 @@
 const postcommentfunc = async (name, comment, itemid) => {
-  console.log('calling postcomment widh: ', name, ' ',comment, ' ', itemid);
   const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/N1YfreMOcnHDHjcZrEgf/comments/', {
     method: 'POST',
     body: JSON.stringify({
@@ -11,7 +10,6 @@ const postcommentfunc = async (name, comment, itemid) => {
       'Content-type': 'application/json',
     },
   });
-  console.log(response.ok);
 }
 
 export default postcommentfunc;
